better directory managment
