+++ CMakeLists.txt
@@ -11,28 +11,10 @@ project(yafth
   LANGUAGES CXX
 )
 
-###
-include(FetchContent)
-
-FetchContent_Declare(ftxui
-  GIT_REPOSITORY https://github.com/ArthurSonzogni/ftxui
-  GIT_TAG v5.0.0
-)
-
-FetchContent_GetProperties(ftxui)
-if(NOT ftxui_POPULATED)
-  FetchContent_Populate(ftxui)
-  add_subdirectory(${ftxui_SOURCE_DIR} ${ftxui_BINARY_DIR} EXCLUDE_FROM_ALL)
-endif()
- ###
- 
-add_executable(yafth src/main.cpp)
-target_include_directories(yafth PRIVATE src)
-
-target_link_libraries(yafth
-  PRIVATE ftxui::screen
-  PRIVATE ftxui::dom
-  PRIVATE ftxui::component
+include_directories(include)
+add_executable(yafth 
+  src/main.cpp 
+  src/yafth_engine.cpp
 )
 
 if(MSVC)
